feat(api)!: add multiple create endpoint
This commit introduces a new POST method to create scale teams in batch.

BREAKING CHANGE: page field changed from Option<u16> to Option<usize> in FtApiProjectSessionsTeamsRequest

+++ libft-api/src/api/project_session/project_sessions_id_teams.rs
@@ -19,7 +19,7 @@ pub struct FtApiProjectSessionsTeamsRequest {
     pub sort: Option<Vec<FtSortOption>>,
     pub filter: Option<Vec<FtFilterOption>>,
     pub range: Option<Vec<FtRangeOption>>,
-    pub page: Option<u16>,
+    pub page: Option<usize>,
     pub per_page: Option<u8>,
 }
